requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
fire==0.5.0
fonttools==4.47.0
fsspec==2023.4.0
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gym==0.26.2
h5py==3.10.0
idna==3.4
imageio==2.33.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.2
medmnist==2.2.3
mpmath==1.3.0
networkx==3.0
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SciencePlots==2.1.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.12.9
torch==2.1.2
torchaudio==2.1.2
torchinfo==1.8.0
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.4
urllib3==2.0.7
validators==0.22.0
Werkzeug==3.0.0